refactored code to conform to new folder structure
