+++ src/components/molecules/loginModal/index.jsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import { useUser } from "../../context/UserContext"
+import { useUser } from "../../../contexts/UserContext"
 
 function LoginModal({ isOpen, handleClose }) {
