+++ src/components/organisms/map/index.jsx
@@ -1,6 +1,6 @@
 import { MapContainer, TileLayer, Marker, Popup } from 'react-leaflet'
 import useSWR from 'swr';
-import { useUser } from '../../context/UserContext';
+import { useUser } from '../../../contexts/UserContext';
 
 const fetcher = async (...args) => await fetch(...args).then(async response => await response.json());
